is_uto/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
networkx>=3.2.1
scipy>=1.12.0
numpy>=1.26.4
ortools>=9.8.3296
folium>=0.16.0
pandas>=2.2.0
openai>=1.30.0
python-dotenv>=1.0.1
pydantic>=2.6.0
